refactor: 랭킹 조회 response 이름 변경
