+++ src/main/java/com/playkuround/playkuroundserver/domain/score/application/LandmarkRankService.java
@@ -2,7 +2,7 @@ package com.playkuround.playkuroundserver.domain.score.application;
 
 import com.playkuround.playkuroundserver.domain.adventure.dao.AdventureRepository;
 import com.playkuround.playkuroundserver.domain.common.DateTimeService;
-import com.playkuround.playkuroundserver.domain.score.api.response.TotalScoreRankingResponse;
+import com.playkuround.playkuroundserver.domain.score.api.response.ScoreRankingResponse;
 import com.playkuround.playkuroundserver.domain.score.dto.NicknameAndScoreAndBadgeType;
 import com.playkuround.playkuroundserver.domain.score.dto.RankAndScore;
 import com.playkuround.playkuroundserver.domain.user.domain.User;
@@ -22,8 +22,8 @@ public class LandmarkRankService {
     private final DateTimeService dateTimeService;
 
     @Transactional(readOnly = true)
-    public TotalScoreRankingResponse getRankTop100ByLandmark(User user, Long landmarkId) {
-        TotalScoreRankingResponse response = TotalScoreRankingResponse.createEmptyResponse();
+    public ScoreRankingResponse getRankTop100ByLandmark(User user, Long landmarkId) {
+        ScoreRankingResponse response = ScoreRankingResponse.createEmptyResponse();
         LocalDateTime monthStartDateTime = DateTimeUtils.getMonthStartDateTime(dateTimeService.getLocalDateNow());
 
         List<NicknameAndScoreAndBadgeType> nicknameAndScores = adventureRepository.findRankTop100DescByLandmarkId(landmarkId, monthStartDateTime);
